requirements.txt
litellm>=1.0.0
gitpython>=3.1.40
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0
pytest>=7.4.0
click>=8.1.0
platformdirs>=4.1.0
sqlalchemy>=2.0.0
